requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.6
django-cachalot==2.6.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
django-tenants==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.12.0
